+++ src/components/banner.astro
@@ -1,6 +1,6 @@
 ---
 import type { SiteConfig } from "@/types/config";
-import { BANNER_HEIGHT_EXTEND } from "@constants/constants";
+import { BANNER_HEIGHT_EXTEND } from "@constants/layout";
 import TypewriterText from "@components/common/typewriterText.astro";
 import ImageWrapper from "@components/common/imageWrapper.astro";
