+++ src/components/sidebar/toc.astro
@@ -1,5 +1,5 @@
 ---
-import { NAVBAR_HEIGHT } from "@constants/constants";
+import { NAVBAR_HEIGHT } from "@constants/layout";
 import { i18n } from "@i18n/translation";
 import I18nKey from "@i18n/i18nKey";
 import WidgetLayout from "./widgetLayout.astro";
