+++ src/components/tocButton.astro
@@ -1,6 +1,6 @@
 ---
 import { Icon } from "astro-icon/components";
-import { NAVBAR_HEIGHT } from "@constants/constants";
+import { NAVBAR_HEIGHT } from "@constants/layout";
 import { i18n } from "@i18n/translation";
 import I18nKey from "@i18n/i18nKey";
