+++ src/constants/category.ts
@@ -1,0 +1,1 @@
+export const UNCATEGORIZED = "uncategorized";
