+++ src/constants/pagination.ts
@@ -1,0 +1,1 @@
+export const PAGE_SIZE = 8;
