+++ src/constants/style.ts
@@ -1,0 +1,3 @@
+export const LIGHT_MODE = "light",
+    DARK_MODE = "dark",
+    SYSTEM_MODE = "system";
