+++ src/pages/projects.astro
@@ -11,7 +11,7 @@ import {
     getFeaturedProjects,
     getAllTechStack,
 } from "@utils/projects";
-import { UNCATEGORIZED } from "@constants/constants";
+import { UNCATEGORIZED } from "@constants/category";
 import { i18n } from "@i18n/translation";
 import I18nKey from "@i18n/i18nKey";
 import ProjectCard from "@components/data/projectCard.astro";
