+++ src/pages/skills.astro
@@ -4,7 +4,7 @@ export const prerender = true;
 
 import { LinkPresets } from "@constants/link-presets";
 import { LinkPreset } from "@/types/config";
-import { UNCATEGORIZED } from "@constants/constants";
+import { UNCATEGORIZED } from "@constants/category";
 import {
     getAdvancedSkills,
     getSkillStats,
